Excising SimplicialComplexes and Tries.
