+++ src/CombinatorialSpaces.jl
@@ -2,7 +2,6 @@ module CombinatorialSpaces
 
 using Reexport
 
-include("Tries.jl")
 include("ArrayUtils.jl")
 include("CombinatorialMaps.jl")
 include("ExteriorCalculus.jl")
@@ -12,14 +11,11 @@ include("MeshInterop.jl")
 include("FastDEC.jl")
 include("Meshes.jl")
 include("restrictions.jl")
-include("SimplicialComplexes.jl")
 include("Multigrid.jl")
 
 
 
-@reexport using .Tries
 @reexport using .SimplicialSets
-@reexport using .SimplicialComplexes
 @reexport using .DiscreteExteriorCalculus
 @reexport using .FastDEC
 @reexport using .MeshInterop
