+++ test/Multigrid.jl
@@ -3,7 +3,7 @@ using Krylov, CombinatorialSpaces, LinearAlgebra, Test
 
 s = triangulated_grid(1,1,1/4,sqrt(3)/2*1/4,Point3D)
 fs = reverse(repeated_subdivisions(4,s,triforce_subdivision_map));
-sses = map(fs) do f dom(f).delta_set end
+sses = map(fs) do f dom(f) end
 push!(sses,s)
 sds = map(sses) do s dualize(s,Circumcenter()) end
 Ls = map(sds) do sd ∇²(0,sd) end
